# www/docs/swagger.yaml
definitions:
  pb.CreateTemplateRequest:
    properties:
      ipId:
        type: integer
      name:
        type: string
    type: object
  pb.GetTemplateRequest:
    properties:
      spotInstanceTemplateId:
        type: integer
    type: object
  pb.LaunchTemplateRequest:
    properties:
      spotInstanceTemplateId:
        type: integer
    type: object
  pb.StopTemplateRequest:
    properties:
      spotInstanceTemplateId:
        type: integer
    type: object
info:
  contact: {}
paths:
  /create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/pb.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create Spot Fleet Template
      tags:
      - Spot
  /get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/pb.GetTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Spot Fleet Instances
      tags:
      - Spot
  /launch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/pb.LaunchTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Launch Spot Fleet Instances
      tags:
      - Spot
  /ping:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /stop:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/pb.StopTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Stop Spot Fleet Instances
      tags:
      - Spot
swagger: "2.0"